#  Improve rectangle operation performance