--- a/hexrd/imageseries/load/__init__.py
+++ b/hexrd/imageseries/load/__init__.py
@@ -1,8 +1,11 @@
 import abc
 import pkgutil
+import numpy as np
 
 from ..imageseriesabc import ImageSeriesABC
 from .registry import Registry
+
+RegionType = tuple[tuple[int, int], tuple[int, int]]
 
 # Metaclass for adapter registry
 
@@ -15,6 +18,13 @@ class _RegisterAdapterClass(abc.ABCMeta):
 class ImageSeriesAdapter(ImageSeriesABC, metaclass=_RegisterAdapterClass):
 
     format = None
+
+    def get_region(self, frame_idx: int, region: RegionType) -> np.ndarray:
+        r = region
+        return self[frame_idx][r[0][0]:r[0][1], r[1][0]:r[1][1]]
+
+    def __getitem__(self, _):
+        pass
 
 # import all adapter modules