--- a/hexrd/utils/hdf5.py
+++ b/hexrd/utils/hdf5.py
@@ -44,6 +44,9 @@ def unwrap_dict_to_h5(grp, d, asattr=False):
                         continue
                     else:
                         # probably a string badness
+
+                        if isinstance(item, np.ndarray) and np.issubdtype(item.dtype, 'U'):
+                            item = str(item) # hdf5 files do not support unicode arrays
                         grp.create_dataset(key, data=item)