# Overview

Provide optimized implementation for when applying the rectangle operator in an imageseries as its first operation. 

ImageSeriesAdapter class has a new method
```
def get_region(self, frame_idx: int, region: RegionType) -> np.ndarray:
    r = region
    return self[frame_idx][r[0][0]:r[0][1], r[1][0]:r[1][1]]
```
Derived classes can offer optimized implementations. 
only ```FrameCacheImageSeriesAdapter```  and ``HDF5ImageSeriesAdapter`` take advantage of this for now.


# Affected Workflows

workflows utilizing `ProcessedImageSeries` should see a  speedup if they use the rectangle operation as their first operation.
Results running locally on a 1441 frame dataset
| format            | before (s) | after(s) | % improvement |
|-------------------|------------|----------|---------------|
| framecache (npz)  | 1020       | 59       | 94%           |
| framecache (fch5) | 935        | 40.3     | 95%           | 
| raw hdf5                | 6020       | 6070     | -0.8 %        | 

# Documentation Changes
no documentation changes